requirements.txt
astroid==3.3.5
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
dill==0.3.9
et_xmlfile==2.0.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
graphviz==0.20.3
httplib2==0.22.0
idna==3.10
isort==5.13.2
mccabe==0.7.0
oauthlib==3.2.2
openpyxl==3.1.5
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==5.28.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pylint==3.3.1
pyparsing==3.2.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
tomlkit==0.13.2
uritemplate==4.1.1
urllib3==2.2.3
